% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancerTypeList.RData}
\alias{cancerTypeList.RData}
\title{List of Cancer Types
\code{cancerTypeList.RData}
Describes a list of the cancer types, i.e. filenames.}
\format{An RData file with 27 filenames/strings:}
\source{
\url{https://www.nature.com/articles/nature12477#mutational-signatures-and-age-of-cancer-diagnosis}
}
\description{
List of Cancer Types
\code{cancerTypeList.RData}
Describes a list of the cancer types, i.e. filenames.
}
\examples{
\dontrun{
load("data/cancerTypeList.RData")
}
}
