% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTransformations.R
\name{convertValidRGB}
\alias{convertValidRGB}
\title{\code{convertValidRGB} converts any "color" object into a valid RGB hex string. Furthermore, it corrects gamma for visibility reasons.}
\usage{
convertValidRGB(colorToConvert, gamma = 1)
}
\arguments{
\item{colorToConvert}{the "color" object}

\item{gamma}{the gamma correction value}
}
\value{
A RGB hex string.
}
\description{
\code{convertValidRGB} converts any "color" object into a valid RGB hex string. Furthermore, it corrects gamma for visibility reasons.
}
\examples{
# Invalid RGB values...
validHex <- convertValidRGB(RGB(1,0,0))
validHex == "#FF0000"
}
