% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawingFlags.R
\name{createAtlas}
\alias{createAtlas}
\title{\code{createAtlas} plots the average flag representation of each specified type. The function determines matching types by prefix matching the column name.}
\usage{
createAtlas(x, types)
}
\arguments{
\item{x}{a flagVisData object, that has nmfData}

\item{types}{the prefix strings that match a column name. An average nmf will be created over all columns that match the string.}
}
\description{
\code{createAtlas} plots the average flag representation of each specified type. The function determines matching types by prefix matching the column name.
}
\examples{
myData <- flagVis(sampleSmallMutationData)
createAtlas(myData, sampleCancerTypeList)
}
\seealso{
\code{\link{createFlag}}
}
