% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawingFlags.R
\name{createFlag}
\alias{createFlag}
\title{\code{createFlag} plots the colors given in a flag like shape}
\usage{
createFlag(flagColors, maxColors = 4, title = "Flag")
}
\arguments{
\item{flagColors}{the colors in the flag, painted from top to bottom. Must be hex strings.}

\item{maxColors}{the maximum number of rows in the flag}

\item{title}{the title at the top of the plot}
}
\description{
\code{createFlag} plots the colors given in a flag like shape
}
\examples{
createFlag(c("#000000", "#FFFFFF", "#00FF00", "#0000FF"))
createFlag(c("#000000", "#FFFFFF", "#00FF00", "#0000FF"), maxColors = 2, title = "Hello!")
}
\seealso{
\code{\link{createAtlas}}
}
