% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\name{flagVis}
\alias{flagVis}
\title{\code{flagVis} reduces the dimensions of 2D matrix data and calculates multiple rgb values representing each sample. The rank of the NMF will be determined by factorsPerColor * numColors.}
\usage{
flagVis(x, factorsPerColor = 2, numColors = 4, seed = 3,
  calculateVarianceRank = FALSE)
}
\arguments{
\item{x}{the 2D dataset. Only non-negative numeric values should be in the matrix. A row cannot contain all zeros.}

\item{factorsPerColor}{the number of factors each color should represent. Currently only supports the value of 2.}

\item{numColors}{the number of colors to output for each column.}

\item{seed}{random seed for repeated runs}

\item{calculateVarianceRank}{option to calculate the ranking of the normalized coefficient matrix rows by standard deviation (1 being highest)}
}
\value{
A flagVisData Class.
}
\description{
\code{flagVis} reduces the dimensions of 2D matrix data and calculates multiple rgb values representing each sample. The rank of the NMF will be determined by factorsPerColor * numColors.
}
\examples{
myData <- flagVis(flagVis::sampleSmallMutationData)
}
