% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\docType{class}
\name{flagVisData-class}
\alias{flagVisData-class}
\title{\code{flagVisData} A class to hold the data returned from the flagVis function. This includes the NMF data, the RGB data, and the normalized coefficient matrix ranking.}
\arguments{
\item{rgbData}{the RGB data, represented as hex strings. Each column may hold multiple RGB hex strings.}

\item{nmfData}{the object returned from the non-negative matrix factorization}

\item{basisOrder}{the ranking of the normalized coefficient matrix rows by standard deviation (1 being highest).}
}
\description{
\code{flagVisData} A class to hold the data returned from the flagVis function. This includes the NMF data, the RGB data, and the normalized coefficient matrix ranking.
}
\examples{
myData <- flagVis(flagVis::sampleSmallMutationData)
createFlag(myData@rgbData[,10])
}
\seealso{
\code{\link{flagVis}}
}
