% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\name{getL1Norms}
\alias{getL1Norms}
\title{\code{getL1Norms} returns the normalization coefficients for each column required for the basis columns. This type of normalization ensures that the column sum is 1}
\usage{
getL1Norms(nmfBasis)
}
\arguments{
\item{nmfBasis}{the basis matrix of the non-negative matrix factorization data}
}
\value{
A numeric vector.
}
\description{
\code{getL1Norms} returns the normalization coefficients for each column required for the basis columns. This type of normalization ensures that the column sum is 1
}
\examples{
x <- getNatureMutatationData(directory)
getL1Norms(basis(nmf(x, 3)))
}
