% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNatureMutationData.R
\name{getNatureMutatationData}
\alias{getNatureMutatationData}
\title{\code{getNatureMutatationData} returns only the filenames with the 96 substitution subtype}
\usage{
getNatureMutatationData(directory)
}
\arguments{
\item{directory}{specifically the directory of the supplementary data from the nature article (https://doi.org/10.1038/nature12477)}
}
\value{
A 2d matrix, with columns representing the cancer subtypes, and the rows representing the mutation frequencies
}
\description{
\code{getNatureMutatationData} returns only the filenames with the 96 substitution subtype
}
\examples{
getNatureMutatationData("~/files/mutational_catalogs/exomes")
}
