% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTransformations.R
\name{getNextColor}
\alias{getNextColor}
\title{\code{getNextColor} internal function for returning different colors based on counter.}
\usage{
getNextColor(i)
}
\arguments{
\item{i}{the counter variable}
}
\value{
A "color" object.
}
\description{
\code{getNextColor} internal function for returning different colors based on counter.
}
