% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\name{getNormalizedCoefSD}
\alias{getNormalizedCoefSD}
\title{\code{getNormalizedCoefSD} calculates the standard deviation for each normalized row of the correlation matrix}
\usage{
getNormalizedCoefSD(nmfResult)
}
\arguments{
\item{nmfResult}{the complete returned data from the non-negative matrix factorization}
}
\value{
A numeric vector.
}
\description{
\code{getNormalizedCoefSD} calculates the standard deviation for each normalized row of the correlation matrix
}
