% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\name{normalizeBasis}
\alias{normalizeBasis}
\title{\code{normalizeBasis} returns the normalized basis, based on the given normalization coefficients. This function divides each column in nmfBasis by the corresponding value in norms.}
\usage{
normalizeBasis(nmfBasis, norms)
}
\arguments{
\item{nmfBasis}{the basis matrix of the non-negative matrix factorization data}

\item{norms}{the normalization coefficients}
}
\value{
A numeric vector.
}
\description{
\code{normalizeBasis} returns the normalized basis, based on the given normalization coefficients. This function divides each column in nmfBasis by the corresponding value in norms.
}
\examples{
  norms <-getL1Norms(basis(nmfResult))
  newBasis <-normalizeBasis(basis(nmfResult), norms)
}
\seealso{
\code{\link{getL1Norms}}
}
