% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagVis.R
\name{normalizeCoef}
\alias{normalizeCoef}
\title{\code{normalizeCoef}  returns the normalized coefficients, based on the given normalization coefficients. This function divides each column in nmfCoef by the corresponding value in norms.}
\usage{
normalizeCoef(nmfCoef, norms)
}
\arguments{
\item{nmfCoef}{the coefficient matrix of the non-negative matrix factorization data}

\item{norms}{the normalization coefficients}
}
\value{
A numeric vector.
}
\description{
\code{normalizeCoef}  returns the normalized coefficients, based on the given normalization coefficients. This function divides each column in nmfCoef by the corresponding value in norms.
}
\examples{
norms <-getL1Norms(basis(nmfResult))
normalizeCoef(coefficients(nmfResult), norms)
}
\seealso{
\code{\link{seq}}
}
