% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNatureMutationData.R
\name{returnExomeFiles}
\alias{returnExomeFiles}
\title{\code{returnExomeFiles} searches through the directory of the Nature Mutation exome data, and returns a list of all readable .txt files}
\usage{
returnExomeFiles(directory)
}
\arguments{
\item{directory}{the directory to search through}
}
\value{
The file strings.
}
\description{
\code{returnExomeFiles} searches through the directory of the Nature Mutation exome data, and returns a list of all readable .txt files
}
\examples{
returnExomeFiles("~/files/mutational_catalogs/exomes")
}
\seealso{
\code{\link{selectSubsitutionSubType}}
}
