% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNatureMutationData.R
\name{selectSubsitutionSubType}
\alias{selectSubsitutionSubType}
\title{\code{selectSubsitutionSubType} returns only the filenames with the 96 substitution subtype}
\usage{
selectSubsitutionSubType(fileList, subtype = "96")
}
\arguments{
\item{fileList}{list of files}

\item{subtype}{the subtype to match. Uses grep pattern matching.}
}
\value{
The file strings.
}
\description{
\code{selectSubsitutionSubType} returns only the filenames with the 96 substitution subtype
}
\seealso{
\code{\link{returnExomeFiles}}
}
