% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTransformations.R
\name{transformPolarLAB}
\alias{transformPolarLAB}
\title{\code{transformPolarLAB} transforms a 2d point into the polarLAB colorspace, given two accent colors. This function first rotates the point 45 degrees, then compresses the space vertically.
Then, the point is scaled in between the hue and lightness of the given accent colors.}
\usage{
transformPolarLAB(x, color1, color2, color1Black = NULL,
  color2Black = NULL)
}
\arguments{
\item{x}{the complete returned data from the non-negative matrix factorization}

\item{color1}{the "color" object for the first accent color, in any representation}

\item{color2}{the "color" object for the second accent color, in any representation}

\item{color1Black}{the optional "color" object, currently only used for the second (darker) lightness. Default value is NULL}

\item{color2Black}{the optional "color" object, currently only used for the second (darker) lightness. Currently ignored. Default value is NULL}
}
\value{
The transformed point as a "color" object in polarLAB space.
}
\description{
\code{transformPolarLAB} transforms a 2d point into the polarLAB colorspace, given two accent colors. This function first rotates the point 45 degrees, then compresses the space vertically.
Then, the point is scaled in between the hue and lightness of the given accent colors.
}
\examples{
myColor <- transformPolarLAB(c(0, 1), COLOR1, COLOR2)
myColorAgain <- transformPolarLAB(c(1, 0), COLOR1, COLOR2,
   colorspace::polarLAB(10,100,260), colorspace::polarLAB(10,100,160))
}
