% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawingFlags.R
\name{createAtlas}
\alias{createAtlas}
\title{\code{createAtlas} plots the average flag representation of each specified type. The function determines matching types by prefix matching the column name.}
\usage{
createAtlas(x, types)
}
\arguments{
\item{x}{a flagVisData object, that has nmfData}

\item{types}{the prefix strings that match a column name. An average nmf will be created over all columns that match the string.}
}
\description{
\code{createAtlas} plots the average flag representation of each specified type. The function determines matching types by prefix matching the column name.
}
\examples{
x <- nmf(x)
myData <- flagVis(x)
files <- returnExomeFiles(directory)
selectedFiles <- selectSubsitutionSubType(files)
for (i in 1:length(selectedFiles)){
   selectedFiles[i] <- tail(strsplit(selectedFiles[i], "/", fixed=TRUE)[[1]], n=1)
}
tail(strsplit(type, "/", fixed=TRUE)[[1]], n=1)
createAtlas(myData, selectedFiles)
}
\seealso{
\code{\link{createFlag}}
}
