% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTransformations.R
\name{getNextColor}
\alias{getNextColor}
\title{\code{getNextColor} internal function for returning different colors based on counter.}
\usage{
getNextColor(i, colorPalette = list(colorspace::RGB(1, 0.1, 0),
  colorspace::RGB(0, 1, 0), colorspace::RGB(0, 0.2, 1), colorspace::RGB(1,
  1, 0), colorspace::RGB(0, 0, 1), colorspace::RGB(0, 1, 0),
  colorspace::RGB(0.5, 0.5, 1), colorspace::RGB(1, 0, 0.5)))
}
\arguments{
\item{i}{the counter variable}

\item{colorPalette}{the color palette, colors repeat once iterated through. Must be of color-class}
}
\value{
A "color" object.
}
\description{
\code{getNextColor} internal function for returning different colors based on counter.
}
